Use per-thread newline count instead of per-tile

+++ include/gpjson/index/kernels/orig.cuh
@@ -1,9 +1,11 @@
+#include <cstddef>
+
 namespace gpjson::index::kernels::orig {
-__global__ void newline_count_index(const char *file, int fileSize,
+__global__ void newline_count_index(const char *file, size_t fileSize,
                                     int *newlineCountIndex);
 
 __global__ void combined_escape_carry_newline_count_index(const char *file,
-                                                          int fileSize,
+                                                          size_t fileSize,
                                                           char *escapeCarry,
                                                           int *newlineCount);
 
@@ -15,22 +17,22 @@ __global__ void int_sum_post_scan(int *intArr, int n, int stride,
 __global__ void int_sum_rebase(int *intArr, int n, int *base, int offset,
                                int *intNewArr);
 
-__global__ void newline_index(const char *file, int fileSize,
+__global__ void newline_index(const char *file, size_t fileSize,
                               int *newlineCountIndex, long *newlineIndex);
 
-__global__ void combined_escape_newline_index(const char *file, int fileSize,
+__global__ void combined_escape_newline_index(const char *file, size_t fileSize,
                                               char *escapeCarryIndex,
                                               int *newlineCountIndex,
                                               long *escapeIndex,
                                               long *newlineIndex);
 
-__global__ void escape_carry_index(const char *file, int fileSize,
+__global__ void escape_carry_index(const char *file, size_t fileSize,
                                    char *escapeCarryIndex);
 
-__global__ void escape_index(const char *file, long fileSize,
+__global__ void escape_index(const char *file, size_t fileSize,
                              char *escapeCarryIndex, long *escapeIndex);
 
-__global__ void quote_index(const char *file, int fileSize, long *escapeIndex,
+__global__ void quote_index(const char *file, size_t fileSize, long *escapeIndex,
                             long *quoteIndex, char *quoteCarryIndex);
 
 __global__ void xor_pre_scan(char *charArr, int n);
@@ -42,7 +44,7 @@ __global__ void xor_rebase(char *charArr, int n, char *base);
 __global__ void string_index(long *stringIndex, int stringIndexSize,
                              char *quoteCarryIndex);
 
-__global__ void leveled_bitmaps_carry_index(const char *file, int fileSize,
+__global__ void leveled_bitmaps_carry_index(const char *file, size_t fileSize,
                                             const long *stringIndex,
                                             char *leveledBitmapsAuxIndex);
 
@@ -54,7 +56,7 @@ __global__ void char_sum_post_scan(char *charArr, int n, int stride,
 __global__ void char_sum_rebase(char *charArr, int n, char *base, int offset,
                                 char *charNewArr);
 
-__global__ void leveled_bitmaps_index(const char *file, int fileSize,
+__global__ void leveled_bitmaps_index(const char *file, size_t fileSize,
                                       const long *stringIndex,
                                       char *leveledBitmapsAuxIndex,
                                       long *leveledBitmapsIndex, int levelSize,
