+++ src/gpjson/index/kernels/sharemem/combined-escape-newline-index.cu
@@ -1,28 +1,29 @@
 #include <cassert>
+#include <cstddef>
 
 namespace gpjson::index::kernels::sharemem {
 
-__global__ void combined_escape_newline_index(const char *file, int fileSize,
+__global__ void combined_escape_newline_index(const char *file, size_t fileSize,
                                               char *escapeCarryIndex,
                                               int *newlineCountIndex,
                                               long *escapeIndex,
                                               long *newlineIndex) {
-  int index = blockIdx.x * blockDim.x + threadIdx.x;
-  int stride = blockDim.x * gridDim.x;
+  size_t index = static_cast<size_t>(blockIdx.x) * blockDim.x + threadIdx.x;
+  size_t stride = static_cast<size_t>(blockDim.x) * gridDim.x;
 
-  int charsPerThread = (fileSize + stride - 1) / stride;
-  int bitmapAlignedCharsPerThread = ((charsPerThread + 64 - 1) / 64) * 64;
-  int start = index * bitmapAlignedCharsPerThread;
-  int end = start + bitmapAlignedCharsPerThread;
+  size_t charsPerThread = (fileSize + stride - 1) / stride;
+  size_t bitmapAlignedCharsPerThread = ((charsPerThread + 64 - 1) / 64) * 64;
+  size_t start = index * bitmapAlignedCharsPerThread;
+  size_t end = start + bitmapAlignedCharsPerThread;
 
   bool carry = index == 0 ? false : escapeCarryIndex[index - 1];
 
   int escapeCount = 0;
-  int totalCount = end - start;
+  size_t totalCount = end - start;
 
   int newlineOffset = newlineCountIndex[index];
 
-  for (long i = start; i < end && i < fileSize; i += 1) {
+  for (size_t i = start; i < end && i < fileSize; i += 1) {
     char value = file[i];
 
     if (carry == 1) {
@@ -37,11 +38,11 @@ __global__ void combined_escape_newline_index(const char *file, int fileSize,
     }
 
     if (value == '\n') {
-      newlineIndex[newlineOffset++] = i;
+      newlineIndex[newlineOffset++] = static_cast<long>(i);
     }
   }
 
-  assert(escapeCount != totalCount);
+  assert(static_cast<size_t>(escapeCount) != totalCount);
 }
 
 } // namespace gpjson::index::kernels::sharemem
